update: extract handleMediaBlocks function to make it reusable
