+++ scripts/utils.js
@@ -39,5 +39,51 @@ export function formatRichText(element, wrapperClassName, textClassName) {
   });
   element.innerHTML = "";
   element.appendChild(wrapper);
-  console.log(wrapper);
+}
+
+/**
+ * Handles media blocks (images/videos) and appends them to section.
+ * @param {NodeListOf<Element>} media
+ * @param {NodeListOf<Element>} imgAltText
+ * @param {HTMLElement} section
+ */
+export function handleMediaBlocks(media, imgAltText, section) {
+  media.forEach((child, idx) => {
+    const isVideo = !child.querySelector("picture");
+
+    if (isVideo) {
+      const anchor = child.querySelector("a");
+      if (anchor) {
+        const video = document.createElement("video");
+        video.autoplay = true;
+        video.playsInline = true;
+        video.muted = true;
+        video.loop = true;
+        video.preload = "none";
+        video.setAttribute("aria-label", "Video player");
+        video.className = "video-player";
+        video.controls = false;
+        const source = document.createElement("source");
+        source.src =
+          "https://publish-p152536-e1620746.adobeaemcloud.com" +
+          anchor.innerText.trim();
+        source.type = "video/webm";
+        video.appendChild(source);
+        anchor.parentElement.remove();
+        child.appendChild(video);
+      }
+    } else {
+      const picture = child.querySelector("picture");
+      const img = picture?.querySelector("img");
+      if (img && imgAltText[idx]) {
+        img.alt = imgAltText[idx].textContent.trim();
+      }
+    }
+
+    if (section) {
+      const childDiv = document.createElement("div");
+      childDiv.appendChild(child);
+      section.appendChild(childDiv);
+    }
+  });
 }
